requirements.txt
keras
h5py
numpy
python-dotenv
boto3
fastapi
uvicorn
pydantic
